Recents: Cancel when new alpha is set before the animation is finished
Dismissing RecentsView to home triggers backgroundScrimAnimator.
And re-entering RecentsActivity sets the alpha value of backgroundScrim
right away without any animation, when there is no task in RecentsView.

So this issue occurs when a user hits home button and backgroundScrimAnimator
gets started, then the user hits recent button right away to enter RecentsActivity again.
Before backgroundScrimAnimator is finished, re-entering RecentsActivity with empty task
instantly sets the alpha of backgroundScrim to 1, then backgroundScrimAnimator sets the alpha to 0 again when it is finished afterward.

This makes backgroundScrim invisible,
and when the wallpaper of the device is something with bright color such as white,
it seems like the whole RecentsView is gone!

Most of all,
there is no need to keep the old animation going on when the alpha value is
already set to the updated value from else where before the animation is over.

Test: Modules build and work.
Change-Id: I3187e69aee7631049e3c4d3305d311eea266fa5d
Signed-off-by: Jinsol Kwon <jinsol.kwon@lge.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: Funny-Bunny <dragos281993@gmail.com>